requirements.txt
addict==2.4.0
einops==0.8.2
gpytorch==1.15.1
mamba_ssm==2.2.6.post3
matplotlib==3.10.8
numpy==2.4.1
opencv_python==4.12.0.88
opencv_python_headless==4.11.0.86
Pillow==12.1.0
psutil==5.9.8
Pympler==1.1
PyYAML==6.0.1
rich==14.3.1
scikit_learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
torch==2.8.0
torchinfo==1.8.0
torchvision==0.23.0
tqdm==4.67.1
ultralytics==8.3.63
ultralytics_thop==2.0.14
